Refactored fhir conversion utils
