+++ src/pages/search.tsx
@@ -8,19 +8,21 @@ import { fhirclient } from 'fhirclient/lib/types';
 import Header from '@/components/Header';
 import PatientCard from '@/components/PatientCard';
 import SearchForm from '@/components/SearchForm';
-import { Patient, convertFhirPatient } from '@/utils/patient';
-import { User, convertFhirUser } from '@/utils/user';
 import {
-  PrimaryCancerCondition,
+  convertFhirEcogPerformanceStatus,
+  convertFhirKarnofskyPerformanceStatus,
+  convertFhirMedicationStatements,
+  convertFhirPatient,
   convertFhirPrimaryCancerCondition,
+  convertFhirRadiationProcedures,
   convertFhirSecondaryCancerCondition,
-} from '@/utils/cancerConditions';
-import { convertFhirEcogPerformanceStatus } from '@/utils/ecogScore';
-import { convertFhirKarnofskyPerformanceStatus } from '@/utils/karnofskyScore';
-import { convertFhirTumorMarkers } from '@/utils/biomarkers';
-import { convertFhirRadiationProcedures } from '@/utils/radiation';
-import { convertFhirSurgeryProcedures } from '@/utils/surgery';
-import { convertFhirMedicationStatements } from '@/utils/medications';
+  convertFhirSurgeryProcedures,
+  convertFhirTumorMarkers,
+  convertFhirUser,
+  Patient,
+  PrimaryCancerCondition,
+  User,
+} from '@/utils/fhirConversionUtils';
 
 type SearchPageProps = {
   patient: Patient;
@@ -55,8 +57,8 @@ const SearchPage = ({
     travelDistance: '100',
     zipcode: patient.zipcode || '',
     metastasis: metastasis || '',
-    ecogScore: ecogScore,
-    karnofskyScore: karnofskyScore,
+    ecogScore,
+    karnofskyScore,
     biomarkers,
     radiation,
     surgery,
@@ -118,7 +120,7 @@ export const getServerSideProps: GetServerSideProps = async context => {
     getMedicationStatement('mcode-cancer-related-medication-statement'),
   ]);
 
-  const serverSideProps = {
+  return {
     props: {
       patient: convertFhirPatient(fhirPatient),
       user: convertFhirUser(fhirUser),
@@ -132,20 +134,15 @@ export const getServerSideProps: GetServerSideProps = async context => {
       medications: convertFhirMedicationStatements(fhirMedicationStatements),
     },
   };
-  console.log('serverSideProps', serverSideProps); // for debugging
-
-  return serverSideProps;
 };
 
 const bundleMaker = (fhirClient: Client) => {
   const urlPatientId = encodeURIComponent(fhirClient.getPatientId());
-  return (resourceType: string) => {
-    return (url: string): Promise<fhirclient.FHIR.Bundle> => {
-      return fhirClient.request<fhirclient.FHIR.Bundle>(
+  return (resourceType: string) =>
+    (url: string): Promise<fhirclient.FHIR.Bundle> =>
+      fhirClient.request<fhirclient.FHIR.Bundle>(
         `${resourceType}?patient=${urlPatientId}&_profile=${encodeURIComponent(
-          `http://hl7.org/fhir/us/mcode/StructureDefinition/` + url
+          `http://hl7.org/fhir/us/mcode/StructureDefinition/${url}`
         )}`
       );
-    };
-  };
 };
